Ajustement des tests avec le token

+++ back-end/API/tests/test_employer.py
@@ -3,6 +3,9 @@ from app import create_app, db
 from app.models.employers_model import Employers
 from app.models.user_model import User
 from app.models.enterprise_model import Enterprise
+from argon2 import PasswordHasher
+
+hasher = PasswordHasher()
 
 @pytest.fixture(scope='module')
 def app():
@@ -24,12 +27,13 @@ def app():
             "enterpriseId": 1,
         }
         employers = Employers(**data)
+        hashed_password = hasher.hash("test")
         db.session.add(employers)
         data = {
             "id": 1,
             "email": "test@test.com",
-            "password": "test",
-            "isModerator": False,
+            "password": hashed_password,
+            "isModerator": True,
             "active": True,
         }
         user = User(**data)
@@ -57,7 +61,13 @@ def test_createEmployer(client):
         "enterpriseId": 1,
         "userId": 1,
     }
-    response = client.post('/employer/createEmployer', json=data)
+    dataLogin = {
+        "email": "test@test.com",
+        "password": "test",
+    }
+    responseLogin = client.post('/user/login', json=dataLogin)
+    token = responseLogin.json['token']
+    response = client.post('/employer/createEmployer', json=data, headers={'Authorization': token})
     assert response.status_code == 200
 
 
@@ -67,14 +77,26 @@ def test_updateEmployer(client):
         "userId": 1,
         "enterpriseId": 1,
     }
-    response = client.put('/employer/updateEmployer/1', json=data)
+    dataLogin = {
+        "email": "test@test.com",
+        "password": "test",
+    }
+    responseLogin = client.post('/user/login', json=dataLogin)
+    token = responseLogin.json['token']
+    response = client.put('/employer/updateEmployer?id=1', json=data,  headers={'Authorization': token})
     assert response.status_code == 200
     assert response.json == {
         "message": "employer updated"
     }
 
 def test_deleteEmployer(client):
-    response = client.delete('/employer/deleteEmployer/1')
+    dataLogin = {
+        "email": "test@test.com",
+        "password": "test",
+    }
+    responseLogin = client.post('/user/login', json=dataLogin)
+    token = responseLogin.json['token']
+    response = client.delete('/employer/deleteEmployer?id=1', headers={'Authorization': token})
     assert response.status_code == 200
     assert response.json == {
         "message": "employer deleted"
